==> final_project/1-predict-ts/1_organize_data.ipynb <==
# Notebook 1: Organizing the dataset

from pathlib import Path
import pandas as pd

def load_data(data_path):   
    operational_settings = ['op_setting_{}'.format(i + 1) for i in range (3)]
    sensor_columns = ['sensor_meas_{}'.format(i + 1) for i in range(26)]
    cols = ['engine_no', 'time_in_cycles'] + operational_settings + sensor_columns
    
    data = pd.read_csv(data_path, sep=' ', header=-1, names=cols)
    data = data.drop(cols[-5:], axis=1)
    data['index'] = data.index
    data.index = data['index']
    data['time'] = pd.date_range('1/1/2000', periods=data.shape[0], freq='600s')
#     print('Loaded data with:\n{} Recordings\n{} Engines'.format(
#         data.shape[0], len(data['engine_no'].unique())))
#     print('21 Sensor Measurements\n3 Operational Settings')
    return data

def load_target(data_path):
    cols = ['RUL']
    data = pd.read_csv(data_path, sep=' ', header=-1, names=cols, index_col=False)
    return data

def make_list(targetx):
    targetx = targetx.values
    target_x=[]
    for item in targetx:
        target_x.append(item[0])
    return target_x

PATH   = Path('../data/CMAPSSData')

# Load dos trains e targets
train01 = load_data(PATH/'train_FD001.txt')
target01 = load_target(PATH/'RUL_FD001.txt')

train02 = load_data(PATH/'train_FD002.txt')
target02 = load_target(PATH/'RUL_FD002.txt')

train03 = load_data(PATH/'train_FD003.txt')
target03 = load_target(PATH/'RUL_FD003.txt')

train04 = load_data(PATH/'train_FD004.txt')
target04 = load_target(PATH/'RUL_FD004.txt')

target02.head(5)

# Arrange engine_no to align with target index
train01['engine_no'] = train01['engine_no'].apply(lambda x: x-1)
train02['engine_no'] = train02['engine_no'].apply(lambda x: x-1)
train03['engine_no'] = train03['engine_no'].apply(lambda x: x-1)
train04['engine_no'] = train04['engine_no'].apply(lambda x: x-1)

# Transformar o target numa lista
target_01= make_list(target01)
target_02= make_list(target02)
target_03= make_list(target03)
target_04= make_list(target04)

# Verificar quais DS precisa dropar a última linha
print("01: ",train01['engine_no'].unique().shape, " ", len(target_01))
print("02: ",train02['engine_no'].unique().shape, " ", len(target_02))
print("03: ",train03['engine_no'].unique().shape, " ", len(target_03))
print("04: ",train04['engine_no'].unique().shape, " ", len(target_04))

# Dropar todoo último engine dos DS 02 
filtered = train02.query('@train02.engine_no == 259')
to_drop = list(filtered['index'].values)

for row in to_drop:
    train02 = train02.drop([row], axis=0)

# Dropar todoo último engine dos DS 02
filtered = train04.query('@train04.engine_no == 248')
to_drop = list(filtered['index'].values)

for row in to_drop:
    train04 = train04.drop([row], axis=0)

# Conferir tamanhos
print("01: ",train01['engine_no'].unique().shape, " ", len(target_01))
print("02: ",train02['engine_no'].unique().shape, " ", len(target_02))
print("03: ",train03['engine_no'].unique().shape, " ", len(target_03))
print("04: ",train04['engine_no'].unique().shape, " ", len(target_04))

# anotar como target correto para cada engine
train01['target'] = train01['engine_no'].apply(lambda x: target_01[x])
train02['target'] = train02['engine_no'].apply(lambda x: target_02[x])
train03['target'] = train03['engine_no'].apply(lambda x: target_03[x])
train04['target'] = train04['engine_no'].apply(lambda x: target_04[x])

# Criar eng_fam

# incluir a coluna com a família do motor
train01['engine_family'] = 1
train02['engine_family'] = 2
train03['engine_family'] = 3
train04['engine_family'] = 4

def make_eng_fam(df):
    df['eng_fam'] = ''
    for i in range(len(df)):
        engine = df.at[i,'engine_no']
        engine = str(int(engine))
        family = str(int(df.at[i, 'engine_family']))
        df.at[i,'eng_fam'] = engine + '_' + family
    return df

train01 = make_eng_fam(train01)

train02 = make_eng_fam(train02)
train03 = make_eng_fam(train03)
train04 = make_eng_fam(train04)

train01.head(5)

# sort by time

train01 = train01.sort_values(by='time')
train02 = train02.sort_values(by='time')
train03 = train03.sort_values(by='time')
train04 = train04.sort_values(by='time')

# Concat all 4 datasets

data = pd.concat([train01, train02, train03, train04], axis = 0, sort=False)

data.isnull().sum()

print(train01.shape)
print(train02.shape)
print(train03.shape)
print(train04.shape)
print(data.shape)

# trocar time por datetime
time = pd.to_datetime(data['time'])
data.drop(['time'], axis=1, inplace=True)
data = pd.concat([data, time], axis=1)

# trocar int por float
for col in ['engine_no', 'time_in_cycles', 'sensor_meas_17','sensor_meas_18', 'engine_family']:
    ints = data[col].astype('float64')
    data.drop([col], axis=1, inplace=True)
    data = pd.concat([data, ints], axis=1)

# persist new dataframe
data.to_hdf('data1.h5', key='data', mode='w')

end =================================================================
