==> final_project/1-predict-ts/2_make_labels.ipynb <==
# Notebook 2: Make Supervised suitable

**Começa em:** Um dataset com tudo  
**Entrega:** X e y

import pandas as pd
import os

data = pd.read_hdf('data1.h5', key='data', parse_dates=['time'])
# data.drop(['Unnamed: 0'], axis=1, inplace=True)

data['target'].describe()

# criar os labels a partir de RUL
data['target'] = data['target'].apply(lambda x: 'falha' if x <= 80 else 'não falha')

data['target'].value_counts()

# Persist
data.to_hdf('data2.h5', key='data', mode='w')

end =================================================================
