==> final_project/1-predict-ts/3_make_3d_keep-Copy1.ipynb <==
# Notebook 3 - Make 3D tensors

import pandas as pd
import numpy as np

data = pd.read_hdf('data2.h5', key='data')

print(data.shape)

data = data.drop(['index', 'engine_no'], axis=1)
data.head(5)

np.amin(size, axis=0)

for i, item in enumerate(cards):
    to_drop = [i for i in range (128, item.shape[0])]
    item.drop(labels=to_drop, axis=0, inplace=True)

# Conferir se deu certo
for item in cards:
    if item.shape[0] is not 128:
        print(item.shape)

# Make 3D array

print(len(cards))
print(type(cards))
print(type(cards[0]))

# teste = np.array(cards, ndmin=3)

# cards_np=np.empty(shape=(707, 128, 29))
# for item in cards:
#     np.insert(obj=cards_np, values=np.array(item.values), axis=3)

df.shape

# concat everything again

df = cards[0].copy(deep=True)
var =1
for item in cards:
        aux = item.copy(deep=True)
        df = pd.concat([df, aux], axis=0)
        var =var+1
        if var%100 is 0:
            print(var, end=' ')

df.shape

df.head(3)

**Separar o target**

y = df[['eng_fam', 'target']].copy(deep=True)

y.shape

# gerar eng_fam unico
# fazer query do target
# gerar único do target
# adicionar na lista
eng_fam_unique = y['eng_fam'].unique()

y.shape

# y.groupby(['eng_fam']).count()

data_3d = np.array(list(df.groupby('eng_fam').apply(pd.DataFrame.as_matrix)))

# https://stackoverflow.com/questions/23478297/pandas-dataframe-or-panel-to-3d-numpy-array

type(data_3d)

data_3d.shape

data_3d[0]

# Persistir o objeto
path="data_3d.npy"
np.save(path, data_3d)

end ====================================================
