==> final_project/1-predict-ts/3_make_3d_keep.ipynb <==
# Notebook 3 - Make 3D tensors

import pandas as pd
import numpy as np

data = pd.read_hdf('data2.h5', key='data')

print(data.shape)

data = data.drop(['index', 'engine_no'], axis=1)
data.head(5)

# Get targets per eng_fam

y = data[['eng_fam', 'target']].copy(deep=True)

data = data.drop(['target'], axis=1)

data.head(3)

data = data.sort_values(by=['eng_fam', 'time'])

# ordenar 
# dar drop no timestamp
data = data.drop(['time'], axis=1)

# Fazer o cards

unique_eng_fam = data['eng_fam'].unique()
len(unique_eng_fam)

unique_eng_fam[:20]

def crop(df):
    print("len df:", len(df))
    to_drop = [i for i in range (128, df.shape[0])]
    df.drop(labels=to_drop, axis=0, inplace=True)

def to_cards(X_df):
    cards = []
    size = []
    for i in range(len(unique_eng_fam)):
        filtered = X_df.query('@X_df.eng_fam == @unique_eng_fam' + str([i]))
        size.append(len(filtered))
        filtered = filtered.reset_index()
#         filtered = filtered.sort_values(['time'])
        filtered = filtered.drop(['index'], axis=1)
#         filtered = filtered.T
        filtered.to_numpy()
        cards.append(filtered)
        
    return cards, size

cards, size = to_cards(data)

for item in size:
    if item is 0:
        print("achei um 0")

print(type(cards))
print(type(cards[0]))

cards[0].head(3)

# Crop cards

np.amin(size, axis=0)

for i, item in enumerate(cards):
    to_drop = [i for i in range (128, item.shape[0])]
    item.drop(labels=to_drop, axis=0, inplace=True)

# Conferir se deu certo
for item in cards:
    if item.shape[0] is not 128:
        print(item.shape)

# Transpose

for i, item in enumerate(cards):
    item = item.T
    cards[i] = item

cards[0].head(5)

# Resize

len(cards)

cards[0].head(1)

item = np.array(cards[0])

item.shape

item

# b = np.resize(a,(2,1,10))
resized = np.resize(item, (28, 1, 128))

resized.shape

resized

for i, item in enumerate(cards):
    item = np.array(item)
    resized = np.resize(item, (28, 1, 128))
    cards[i] = resized

cards[0].shape

np.save('data_3d', cards)

# Separar o target

y.shape

y.head(4)

y_g = y.groupby(['eng_fam'])

y_true = y_g.head(1)

y_values = y_true.values

y_pure = []
for i, item in enumerate(y_values):
#     print(item)
    y_pure.append(y_values[i][1])

len(y_pure)

y_pure = np.array(y_pure)

y_pure[:10]

np.save('y', y_pure)

end ====================================================
