==> final_project/1-predict-ts/4_cnn.ipynb <==
import pandas as pd
import numpy as np
from keras import Sequential
from keras.layers import Conv2D, MaxPooling2D, Activation, Dense,Flatten, Dropout
from keras.optimizers import Adam
from sklearn.model_selection import train_test_split

X = np.load("data_3d.npy", allow_pickle=True)
y = np.load("y.npy", allow_pickle=True)

y.shape

y = y.squeeze()

y = pd.get_dummies(y)

y.head(3)

n_classes=2
# input_shape = (1, 128, 28)
input_shape = (28, 1, 128)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

model = Sequential()
model.add(Conv2D(filters=128, kernel_size=(5, 5), activation="relu", padding="same", name="Conv1", input_shape=input_shape))
model.add(Conv2D(filters=64, kernel_size=(5, 5), activation="relu", padding="same", name="Conv2"))
# model.add(Conv2D(filters=32, kernel_size=(5, 5), activation="relu", padding="same", name="Conv3"))
# model.add(Conv2D(filters=16, kernel_size=(5, 5), activation="relu", padding="same", name="Conv4"))

# model.add(MaxPooling2D(pool_size=(2, 2), name="Pool1"))
model.add(Flatten(name="Flatten"))
model.add(Dense(128, activation="relu", name="Dense1"))
model.add(Dropout(0.4, name="Dropout"))
model.add(Dense(128, activation="relu", name="Dense2"))
model.add(Dropout(0.4, name="Dropout2"))
model.add(Dense(n_classes, activation='softmax', name="Output"))

# Resumo do modelo
model.summary()

# opt = Adam(lr=1e-4, decay=1e-4 / args["epochs"])
model.compile(loss="binary_crossentropy", optimizer='adam', metrics=["accuracy"])

history = model.fit(X_train, y_train, batch_size=128, epochs=100, validation_data=(X_test, y_test))
